src/pyproject.toml
[tool.poetry]
name = "MedicalImaging"
version = "0.1.0"
description = ""
authors = ["戌亥 <mi69002@e.hiroshima-cu.ac.jp>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.10.0"
pillow = "^10.3.0"
matplotlib = "^3.9.0"
numpy = "^1.26.4"
rembg = "^2.0.57"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
